Adds control of tp-link smart plug
Sends JSON formatted string to smart outlet using socket
connection, after properly encrypting. The answer is
decrypted, to check if the command was properly accepted.
Functions for turning the switch on/off are included, as
well as checking the plugs current state.

Basic testing is done in loop by flipping state
(on/off) every 5 seconds.