Tweaks polling delays and adds README instructions
Previously the ESP8266 only polled the Sonos every 10 seconds,
and checked if the plug was on or off every time.
This negatively affects how long it takes the system to react
when a trigger event occurs. The plug state (on/off) is no
longer checked since the commands are distinct ON and OFF (no toggle).
The poll time is also reduced to 1 second to react more quickly
to changes in Sonos state.

Previously no instructions were included, now basic instructions
are.