Re-connect wifi when Sonos can't connect, check for line-in
Sometimes the Sonos can't be reached, which causes the plug to turn
off even if music is playing. This is now handled by checking the
Wifi connection, and resetting the time since music was last heard.
As long as a connection with the Sonos is re-established within a
reasonable amount of time the amp shouldn't stay powered on
unnecessarily.

When the source is LINEIN, the Sonos isn't playing anything so the
plug turns off within 5 minutes. For now the plug will never be
turned off if LINEIN is selected as the source. Hopefully a better
solution (querying the Sonos for activity on LINEIN? currently it
doesn't seem this is available) will show up in the future.